사용자에따른 UI구현

+++ 20210828/src/bbs/BbsApp.java
@@ -11,112 +11,170 @@ public class BbsApp {
 		int ls = 0; // 0:로그인안한상태	 	1:회원로그인상태 		2:관리자로그인상태
 		int selectMenu;
 		while(true) {
-			try {
-				if(ls == 0) {
-					System.out.println("-------------------------------------------------------------------------------------------------------------");
-				} else {		//로그인했을때만 왼쪽 상단에 아이디표시
-					System.out.println(bbsDao.userId + "--------------------------------------------------------------------------------------------------------");
-				}
-				System.out.println("0.로그인   1.글쓰기   2.글찾기   3.글수정   4.글삭제(관리자) 5.글전체 조회   6.회원가입   7.로그아웃  8.회원삭제(관리자) 9.회원목록조회(관리자)  10.종료");
-				System.out.println("-------------------------------------------------------------------------------------------------------------");
-				selectMenu = sc.nextInt();
-				sc.nextLine();
-				if(selectMenu < MENU.LOGIN || selectMenu > MENU.EXIT)
-					throw new MenuException(selectMenu);
-				switch(selectMenu) {
-				case MENU.LOGIN : 
-					if(ls == 1 || ls == 2) {
-						System.out.println("이미 로그인상태 입니다."); 
-						break;
-					} else {
-						ls = loginBbs(); break;
+			if(ls != 0) {
+				System.out.println("---------------");
+				System.out.println(bbsDao.userId);
+			}
+			if(ls == 0) {			//로그인 안했을때 메뉴
+				try {
+	//				System.out.println("---------------------------------------------------------------------------------------------------------------------");
+	//				System.out.println("0.로그인   1.글쓰기   2.글찾기   3.글수정   4.글삭제(관리자) 5.글전체 조회   6.회원가입   7.로그아웃  8.회원삭제(관리자) 9.회원목록조회(관리자)  10.종료");
+	//				System.out.println("---------------------------------------------------------------------------------------------------------------------");
+					System.out.println("---------------------------------------------------------------------------------------------------------------------");
+					System.out.println("0.로그인   1.글찾기   2.글전체 조회   3.회원가입   4.종료");
+					System.out.println("---------------------------------------------------------------------------------------------------------------------");
+					selectMenu = sc.nextInt();
+					sc.nextLine();
+					if(selectMenu < MENU.LOGIN || selectMenu > MENU.EXIT)
+						throw new MenuException(selectMenu);
+					switch(selectMenu) {
+					case MENU.LOGIN : 
+//						if(ls == 1 || ls == 2) {
+//							System.out.println("이미 로그인상태 입니다."); 
+//							break;
+//						} else {
+							ls = loginBbs(); break;
+//						}
+	//				case MENU.INSERT:
+	//					if(ls == 1 || ls == 2) {
+	//						insertBbs(); break;
+	//					} else {
+	//						System.out.println("로그인을 해주세요.");
+	//						break;
+	//					}
+					case MENU.SEARCH:
+						searchBbs(); break;
+	//				case MENU.UPDATE:
+	//					if(ls == 1 || ls == 2) {
+	//					updateBbs(); break;
+	//					} else {
+	//						System.out.println("로그인 해주세요.");
+	//						break;
+	//					}
+	//				case MENU.DELETE:
+	//					if(ls == 2) {
+	//						deleteBbsMember(); break;
+	//					} else {
+	//						System.out.println("권한이 없습니다.");
+	//						System.out.println("관리자로 로그인 해주세요.");
+	//						break;
+	//					}
+					case MENU.SEARCHALL:
+						searchAllBbs(); break;
+					case MENU.JOIN:
+						joinBbs(); break;
+	//				case MENU.LOGOUT:
+	//					if(ls == 1 || ls == 2) {
+	//						ls = 0;
+	//						System.out.println("로그아웃하였습니다.");
+	//						break;
+	//					} else {
+	//						System.out.println("로그인이 되어있지 않습니다.");
+	//						break;
+	//					}
+	//				case MENU.DELETEMEMBER:
+	//					if(ls == 2) {
+	//						deleteBbsMember(); break;
+	//					} else {
+	//						System.out.println("권한이 없습니다.");
+	//						System.out.println("관리자로 로그인 해주세요.");
+	//						break;
+	//					}
+	//				case MENU.MEMBERLIST:
+	//					if(ls == 2) {
+	//						selectMemberList(); break;
+	//					} else {
+	//						System.out.println("권한이 없습니다.");
+	//						System.out.println("관리자로 로그인 해주세요.");
+	//						break;
+	//					}
+					case MENU.EXIT:
+						System.out.println("종료");
+						return;
 					}
-				case MENU.INSERT:
-					if(ls == 1 || ls == 2) {
+				} catch (MenuException e) {
+					e.showWrongMessage();
+				}
+			}
+			else if(ls == 1) {			//회원로그인 했을때 메뉴
+				try {
+					System.out.println("---------------------------------------------------------------------------------------------------------------------");
+					System.out.println("0.글쓰기   1.글찾기   2.글수정   3.글삭제   4.글전체 조회   5.로그아웃   6.종료");
+					System.out.println("---------------------------------------------------------------------------------------------------------------------");
+					selectMenu = sc.nextInt();
+					sc.nextLine();
+					if(selectMenu < MemberMENU.INSERT || selectMenu > MemberMENU.EXIT)
+						throw new MenuException(selectMenu);
+					switch(selectMenu) {
+					case MemberMENU.INSERT:
 						insertBbs(); break;
-					} else {
-						System.out.println("로그인을 해주세요.");
-						break;
-					}
-				case MENU.SEARCH:
-					searchBbs(); break;
-				case MENU.UPDATE:
-					if(ls == 1 || ls == 2) {
-					updateBbs(); break;
-					} else {
-						System.out.println("로그인 해주세요.");
-						break;
-					}
-				case MENU.DELETE:
-					if(ls == 2) {
-						deleteBbsMember(); break;
-					} else {
-						System.out.println("권한이 없습니다.");
-						System.out.println("관리자로 로그인 해주세요.");
-						break;
-					}
-				case MENU.SEARCHALL:
-					searchAllBbs(); break;
-				case MENU.JOIN:
-					joinBbs(); break;
-				case MENU.LOGOUT:
-					if(ls == 1 || ls == 2) {
+					case MemberMENU.SEARCH:
+						searchBbs(); break;
+					case MemberMENU.UPDATE:
+						updateBbs(); break;
+					case MemberMENU.DELETE:
+						deleteBbs(ls); break;
+					case MemberMENU.SEARCHALL:
+						searchAllBbs(); break;
+					case MemberMENU.LOGOUT:
 						ls = 0;
-						System.out.println("로그아웃하였습니다.");
+						System.out.println("로그아웃 하였습니다.");
 						break;
-					} else {
-						System.out.println("로그인이 되어있지 않습니다.");
-						break;
+					case MemberMENU.EXIT:
+						System.out.println("종료");
+						return;
 					}
-				case MENU.DELETEMEMBER:
-					if(ls == 2) {
-						deleteBbsMember(); break;
-					} else {
-						System.out.println("권한이 없습니다.");
-						System.out.println("관리자로 로그인 해주세요.");
-						break;
-					}
-				case MENU.MEMBERLIST:
-//					onlyAdmin(ls); break;
-					if(ls == 2) {
-						selectMemberList(); break;
-					} else {
-						System.out.println("권한이 없습니다.");
-						System.out.println("관리자로 로그인 해주세요.");
-						break;
-					}
-				case MENU.EXIT:
-					System.out.println("종료");
-					return;
+				} catch (MenuException e) {
+					e.showWrongMessage();
 				}
-			} catch (MenuException e) {
-				e.showWrongMessage();
+			}
+			else {			//관리자 로그인 했을때 메뉴
+				try {
+					System.out.println("---------------------------------------------------------------------------------------------------------------------");
+					System.out.println("0.글쓰기   1.글찾기   2.글수정   3.글삭제   4.글전체 조회   5.회원삭제   6.회원목록   7.로그아웃  8.종료");
+					System.out.println("---------------------------------------------------------------------------------------------------------------------");
+					selectMenu = sc.nextInt();
+					sc.nextLine();
+					if(selectMenu < AdminMENU.INSERT || selectMenu > AdminMENU.EXIT)
+						throw new MenuException(selectMenu);
+					switch(selectMenu) {
+					case AdminMENU.INSERT:
+							insertBbs(); break;
+					case AdminMENU.SEARCH:
+						searchBbs(); break;
+					case AdminMENU.UPDATE:
+						updateBbs(); break;
+					case AdminMENU.DELETE:
+						deleteBbs(ls); break;
+					case AdminMENU.SEARCHALL:
+						searchAllBbs(); break;
+					case AdminMENU.DELETEMEMBER:
+						deleteBbsMember(); break;
+					case AdminMENU.MEMBERLIST:
+						selectMemberList(); break;
+					case AdminMENU.LOGOUT:
+						ls = 0;
+						System.out.println("로그아웃 하였습니다.");
+						break;
+					case AdminMENU.EXIT:
+						System.out.println("종료");
+						return;
+					}
+				} catch (MenuException e) {
+					e.showWrongMessage();
+				}
 			}
 		}
 		
 		
 	}
 	
-	//관리자권한경고
-//	static void onlyAdmin(int ls) {
-//		if(ls == 2) {
-//			deleteBbsMember();
-//		} else {
-//			System.out.println("권한이 없습니다."); 
-//			System.out.println("관리자로 로그인 해주세요.");
-//		}
-//	}
-	
 	static void insertBbs() {
 		BbsVO bbs = new BbsVO();
 		System.out.println("-----------------------------글쓰기-----------------------------");
-//		System.out.print("작성자 : ");
-//		String name = sc.nextLine();
 		System.out.print("제목 : ");
 		String title = sc.nextLine();
-//		System.out.println("내용 : ");
 		String content = ScannerUtil.readMultiLine();
-//		bbs.setName(name);
 		bbs.setTitle(title);
 		bbs.setContent(content);
 		bbsDao.insert(bbs);
@@ -127,6 +185,7 @@ public class BbsApp {
 		System.out.println("작성자로 글 찾기");
 		System.out.print("작성자 : ");
 		String name = sc.nextLine();
+		System.out.println("-----------------------------글목록-----------------------------");
 		
 		bbsDao.search(name);
 	}
@@ -145,8 +204,6 @@ public class BbsApp {
 			reTitle = bbsDao.getBbs(title).getTitle();
 		}
 		
-//		System.out.println("내용");
-//		String content = sc.nextLine();
 		String content = ScannerUtil.readMultiLine();
 		if(!content.equals("")) {
 			bbs.setContent(content);
@@ -157,11 +214,15 @@ public class BbsApp {
 		bbsDao.update(reTitle, content, title);
 	}
 	
-	static void deleteBbs() {
+	static void deleteBbs(int ls) {
 		System.out.println("-----------------------------글삭제-----------------------------");
 		System.out.print("삭제할 글의 제목 : ");
 		String title = sc.nextLine();
-		bbsDao.delete(title);
+		if(ls == 1) {
+			bbsDao.delete(title);
+		} else {
+			bbsDao.adminDelete(title);
+		}
 	}
 	
 	static void searchAllBbs() {
